Move auth/github to git/github
