+++ git/credentials.go
@@ -22,7 +22,7 @@ import (
 	"time"
 
 	"github.com/fluxcd/pkg/auth/azure"
-	"github.com/fluxcd/pkg/auth/github"
+	"github.com/fluxcd/pkg/git/github"
 )
 
 const (
